Rename DepartmentRoutes and fix imports
